+++ src/autoskillit/server/_recipe_execution.py
@@ -2,14 +2,13 @@
 
 from __future__ import annotations
 
-import hashlib
 import json
 import threading
 from collections.abc import Mapping
 from dataclasses import replace
 from pathlib import Path
 from types import MappingProxyType
-from typing import TYPE_CHECKING, Any
+from typing import TYPE_CHECKING
 from uuid import uuid4
 
 from autoskillit.core import (
@@ -21,7 +20,6 @@ from autoskillit.core import (
     AuditVerdict,
     BindingMode,
     BoundScalar,
-    BoundValueState,
     InstalledRecipeExecution,
     InventoryAdmissionDecision,
     InvocationTemplate,
@@ -36,11 +34,13 @@ from autoskillit.core import (
     compute_tool_contract_identity,
     get_logger,
     get_tool_def,
-    resolve_skill_name,
 )
 from autoskillit.recipe import (
     RecipeStep,
+    RuntimeBindingError,
+    bind_runtime_skill_invocation,
     bind_step_invocation,
+    compute_skill_contract_identity,
     get_skill_contract,
     load_bundled_manifest,
 )
@@ -67,8 +67,6 @@ __all__ = [
     "publish_verified_audit_cycle",
     "record_runtime_binding_digest",
 ]
-
-_SKILL_CONTRACT_IDENTITY_DOMAIN = b"autoskillit:skill-contract:v1\0"
 
 
 class AuditCycleHeadConflict(RuntimeError):
@@ -282,37 +280,6 @@ class DefaultInputPreflightResolver:
         return self._result(decision)
 
 
-def _skill_contract_identity(
-    skill_name: str,
-    *,
-    manifest: dict[str, Any] | None = None,
-) -> str:
-    active_manifest = manifest if manifest is not None else load_bundled_manifest()
-    contract = get_skill_contract(skill_name, active_manifest)
-    if contract is None:
-        raise ValueError(f"skill contract is unavailable for {skill_name!r}")
-    payload = json.dumps(
-        {
-            "completion_required": contract.completion_required,
-            "input_preflight": contract.input_preflight,
-            "inputs": [
-                {
-                    "name": item.name,
-                    "nullable": item.nullable,
-                    "required": item.required,
-                    "type": item.type,
-                }
-                for item in contract.inputs
-            ],
-            "skill_name": skill_name,
-        },
-        ensure_ascii=False,
-        sort_keys=True,
-        separators=(",", ":"),
-    ).encode("utf-8")
-    return "sha256:" + hashlib.sha256(_SKILL_CONTRACT_IDENTITY_DOMAIN + payload).hexdigest()
-
-
 def _run_skill_tool_contract_identity() -> str:
     tool_def = get_tool_def("run_skill")
     if tool_def is None:
@@ -337,7 +304,7 @@ def build_recipe_execution_snapshot(
             continue
         if invocation.skill_name is None:
             raise AssertionError("attested invocation is missing its skill identity")
-        skill_identity = _skill_contract_identity(invocation.skill_name)
+        skill_identity = compute_skill_contract_identity(invocation.skill_name)
         digest = compute_invocation_template_digest(
             execution_id=active_execution_id,
             recipe_name=recipe_name,
@@ -449,10 +416,6 @@ def record_runtime_binding_digest(
         )
 
 
-def _is_dynamic(value: Any) -> bool:
-    return bool(value.context_dependencies or value.input_dependencies)
-
-
 def bind_attested_runtime_invocation(
     installed: InstalledRecipeExecution,
     *,
@@ -481,131 +444,18 @@ def bind_attested_runtime_invocation(
             "invocation_template_digest_mismatch",
             "invocation template digest does not match the active step",
         )
-    invocation = template.invocation
-    if resolve_skill_name(skill_command) != invocation.skill_name:
-        raise RecipeExecutionAdmissionError(
-            "recipe_execution_skill_mismatch",
-            "runtime skill identity differs from the compiled template",
+    try:
+        bound_inputs = bind_runtime_skill_invocation(
+            template,
+            execution_id=execution_id,
+            step_name=step_name,
+            skill_command=skill_command,
+            skill_inputs=skill_inputs,
+            actual_mcp_kwargs=actual_mcp_kwargs,
         )
-    compiled_mcp_names = frozenset(value.name for value in invocation.mcp_kwargs)
-    protocol_mcp_values = {
-        "step_name": step_name,
-        "recipe_execution_id": execution_id,
-        "invocation_template_digest": template_digest,
-    }
-    for name, expected in protocol_mcp_values.items():
-        if name in actual_mcp_kwargs and actual_mcp_kwargs[name] != expected:
-            raise RecipeExecutionAdmissionError(
-                "recipe_execution_tool_shape",
-                f"attestation parameter {name!r} differs from the active invocation",
-            )
-    undeclared_effective_names = sorted(
-        name
-        for name, value in actual_mcp_kwargs.items()
-        if name not in compiled_mcp_names and name not in protocol_mcp_values and value != ""
-    )
-    if undeclared_effective_names:
-        raise RecipeExecutionAdmissionError(
-            "recipe_execution_tool_shape",
-            (
-                "runtime tool parameters are absent from the compiled template: "
-                f"{undeclared_effective_names!r}"
-            ),
-        )
-    manifest = load_bundled_manifest()
-    contract = get_skill_contract(invocation.skill_name or "", manifest)
-    if contract is None:
-        raise RecipeExecutionAdmissionError(
-            "recipe_execution_contract_unavailable",
-            "the compiled skill contract is unavailable at runtime",
-        )
-    runtime_skill_identity = _skill_contract_identity(
-        invocation.skill_name or "",
-        manifest=manifest,
-    )
-    if runtime_skill_identity != template.skill_contract_identity:
-        raise RecipeExecutionAdmissionError(
-            "recipe_execution_contract_mismatch",
-            "runtime skill contract differs from the compiled template",
-        )
-    contract_inputs = {input_def.name: input_def for input_def in contract.inputs}
-    supplied = dict(skill_inputs or {})
-    if skill_inputs is None:
-        runtime_with_args: dict[str, object] = {"skill_command": skill_command}
-        runtime_cwd = actual_mcp_kwargs.get("cwd")
-        if isinstance(runtime_cwd, str):
-            runtime_with_args["cwd"] = runtime_cwd
-        runtime_inline = bind_step_invocation(
-            step_name,
-            RecipeStep(
-                name=step_name,
-                tool="run_skill",
-                with_args=runtime_with_args,
-                declared_with_args=dict(runtime_with_args),
-            ),
-            manifest=manifest,
-        )
-        if runtime_inline.failures:
-            raise RecipeExecutionAdmissionError(
-                "recipe_execution_input_shape",
-                runtime_inline.failures[0].message,
-            )
-        supplied = {
-            value.name: value.effective_value
-            for value in runtime_inline.skill_inputs
-            if value.state is BoundValueState.PRESENT
-            and isinstance(value.effective_value, (str, int, bool))
-        }
-    if any(
-        not isinstance(value, (str, int, bool)) or value is None for value in supplied.values()
-    ):
-        raise RecipeExecutionAdmissionError(
-            "recipe_execution_input_type",
-            "skill_inputs values must be strict JSON scalars",
-        )
-    expected_names = tuple(
-        value.name for value in invocation.skill_inputs if value.state is BoundValueState.PRESENT
-    )
-    if frozenset(supplied) != frozenset(expected_names):
-        raise RecipeExecutionAdmissionError(
-            "recipe_execution_input_shape",
-            "skill_inputs keys do not exactly match the compiled template",
-        )
-    bound_inputs: list[tuple[str, BoundScalar]] = []
-    for value in invocation.skill_inputs:
-        if value.state is not BoundValueState.PRESENT:
-            continue
-        actual = supplied[value.name]
-        input_def = contract_inputs.get(value.name)
-        if input_def is None:
-            raise RecipeExecutionAdmissionError(
-                "recipe_execution_contract_mismatch",
-                f"compiled skill input {value.name!r} is absent from the runtime contract",
-            )
-        if not input_def.accepts(actual):
-            raise RecipeExecutionAdmissionError(
-                "recipe_execution_input_type",
-                f"runtime skill input {value.name!r} expects {input_def.type!r}",
-            )
-        if not _is_dynamic(value) and actual != value.effective_value:
-            raise RecipeExecutionAdmissionError(
-                "recipe_execution_static_input_mismatch",
-                f"static skill input {value.name!r} differs from the template",
-            )
-        bound_inputs.append((value.name, actual))
-    for value in invocation.mcp_kwargs:
-        if value.name not in actual_mcp_kwargs:
-            raise RecipeExecutionAdmissionError(
-                "recipe_execution_tool_shape",
-                f"compiled tool parameter {value.name!r} is absent",
-            )
-        actual = actual_mcp_kwargs[value.name]
-        if not _is_dynamic(value) and actual != value.effective_value:
-            raise RecipeExecutionAdmissionError(
-                "recipe_execution_static_tool_mismatch",
-                f"static tool parameter {value.name!r} differs from the template",
-            )
-    return tuple(bound_inputs), template
+    except RuntimeBindingError as exc:
+        raise RecipeExecutionAdmissionError(exc.code, str(exc)) from exc
+    return bound_inputs, template
 
 
 def build_bound_child_prompt(
